Subclass nova/compute/api
