+++ chef/cookbooks/bcpc/files/default/nova/server_system_metadata_api.py
@@ -21,7 +21,7 @@ from nova.api.openstack import common
 from nova.api.openstack.compute.schemas import server_system_metadata
 from nova.api.openstack import wsgi
 from nova.api import validation
-from nova.compute import api as compute
+from nova.compute.bcpc import api as compute
 from nova import exception
 from nova.i18n import _
 from nova.policies import server_system_metadata as ssm_policies
